Move wrapper headers to a separate include dir
