+++ core/arch/arch.h
@@ -44,7 +44,7 @@
 #ifndef ARCH_H
 #define ARCH_H
 
-#include "../stddef_wrapper.h" /* for offsetof */
+#include "stddef_wrapper.h" /* for offsetof */
 #include "instr.h"             /* for reg_id_t */
 #include "decode.h"            /* for X64_CACHE_MODE_DC */
 #include "arch_exports.h"      /* for FRAG_IS_32 and FRAG_IS_X86_TO_X64 */
