+++ core/include/stdarg_wrapper.h
