+++ core/lib/globals_shared.h
@@ -76,7 +76,7 @@
 
 #include "globals_api.h" // IWYU pragma: export
 
-#include "../limits_wrapper.h" /* for USHRT_MAX */
+#include "limits_wrapper.h" /* for USHRT_MAX */
 
 #ifdef UNIX
 #    ifdef LINUX_KERNEL
